+++ disparity_view/view.py
@@ -1,7 +1,7 @@
 """
 library to view disparity npy files.
 """
-
+import argparse
 import time
 from pathlib import Path
 from typing import Tuple
@@ -191,3 +191,29 @@ def disparity_viewer_main():
         view3d(args)
     else:
         view_by_colormap(args)
+
+
+def view_npy_main():
+    parser = argparse.ArgumentParser(description="np file viewer")
+    parser.add_argument("npy_file", help="npy_file to view")
+    parser.add_argument("--vmax", type=float, default=500, help="max disparity [pixel]")
+    parser.add_argument("--vmin", type=float, default=0, help="min disparity [pixel]")
+    parser.add_argument("--disp3d", action="store_true", help="display 3D")
+    parser.add_argument("--save", action="store_true", help="save colored or ply")
+    group = parser.add_argument_group("colormap")
+    group.add_argument("--gray", action="store_true", help="gray colormap")
+    group.add_argument("--jet", action="store_true", help="jet colormap")
+    group.add_argument("--inferno", action="store_true", help="inferno colormap")
+
+    args = parser.parse_args()
+    print(args)
+    if Path(args.npy_file).is_file():
+        disparity = np.load(args.npy_file)
+        view_npy(disparity, args)
+    elif Path(args.npy_file).is_dir():
+        npys = sorted(Path(args.npy_file).glob("*.npy"))
+        for npy in tqdm(npys):
+            disparity = np.load(npy)
+            view_npy(disparity, args)
+    else:
+        print(f"no such file {args.npy_file}")
